+++ rychkov.mihail/common/math.hpp
@@ -3,7 +3,7 @@
 
 namespace rychkov
 {
-  bool isAlmostEqual(double a, double b, int coef);
+  bool isAlmostEqual(double a, double b, double coef);
   double getEpsilon(double num, bool toPlusInfinity);
   constexpr double PI = 3.14159265358979323846;
 }
